python/requirements.txt
argh==0.26.2
asn1crypto==0.24.0
attrs==19.3.0
autobahn==17.10.1
Automat==0.6.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.4
cbor==1.0.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==2.1.4
cycler==0.10.0
Cython==0.26.1
decorator==4.4.2
defusedxml==0.6.0
dtrx==7.1
entrypoints==0.3
enum34==1.1.10
functools32==3.2.3.post2
futures==3.3.0
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.6.0
incremental==16.10.1
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.3
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
lz4==0.10.1
MarkupSafe==1.1.1
matplotlib==2.2.5
meld3==1.0.2
mistune==0.8.4
mpi4py==2.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.6
olefile==0.45.1
PAM==0.4.2
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.5
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
py-ubjson==0.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygmaps-enhanced==0.2.0
Pygments==2.5.2
pygobject==3.26.1
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.4
pyshp==2.1.2
python-dateutil==2.8.1
python-snappy==0.5
PyTrie==0.2
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==19.0.0
qrcode==5.3
sanic-admin==0.0.6
scandir==1.10.0
scipy==1.2.3
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
subprocess32==3.5.4
supervisor==3.3.1
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
trollius==2.0.1
Twisted==17.9.0
txaio==2.8.1
u-msgpack-python==2.1
watchdog==0.9.0
wcwidth==0.1.9
webencodings==0.5.1
wsaccel==0.6.2
xgboost==0.82
zipp==1.2.0
zope.interface==4.3.2
